requirements.txt
google-adk[extensions]
anthropic>=0.43.0
python-dotenv
httpx
psycopg2-binary